+++ tests/test.lunaris.storeAndHook.js
@@ -194,7 +194,7 @@ describe('lunaris store', function () {
         should(lastError.length).eql(2);
         should(lastError[0]).eql('[Lunaris warn] lunaris.insert@store_insert_post');
         should(lastError[1]).eql({ error : 404, message : 'Not Found'});
-        should(err).eql('404 : Not Found');
+        should(err).eql('${An error has occured}');
         done();
       });
 
@@ -219,7 +219,7 @@ describe('lunaris store', function () {
           url                : '/store_insert_post',
           method             : 'POST',
           storeName          : 'store_insert_post',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -312,7 +312,7 @@ describe('lunaris store', function () {
         should(lastError.length).eql(2);
         should(lastError[0]).eql('[Lunaris warn] lunaris.update@store_insert_put');
         should(lastError[1]).eql({ error : 404, message : 'Not Found'});
-        should(err).eql('404 : Not Found');
+        should(err).eql('${An error has occured}');
         done();
       });
 
@@ -377,7 +377,7 @@ describe('lunaris store', function () {
           url                : '/store_insert_put/1',
           method             : 'PUT',
           storeName          : 'store_insert_put',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -723,7 +723,7 @@ describe('lunaris store', function () {
         should(lastError.length).eql(2);
         should(lastError[0]).eql('[Lunaris warn] lunaris.insert@store_insert_post');
         should(lastError[1]).eql({ error : 404, message : 'Not Found'});
-        should(err).eql('404 : Not Found');
+        should(err).eql('${An error has occured}');
         done();
       });
 
@@ -748,7 +748,7 @@ describe('lunaris store', function () {
           url                : '/store_insert_post',
           method             : 'POST',
           storeName          : 'store_insert_post',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -869,7 +869,7 @@ describe('lunaris store', function () {
         should(lastError.length).eql(2);
         should(lastError[0]).eql('[Lunaris warn] lunaris.update@store_insert_put');
         should(lastError[1]).eql({ error : 404, message : 'Not Found'});
-        should(err).eql('404 : Not Found');
+        should(err).eql('${An error has occured}');
         done();
       });
 
@@ -930,7 +930,7 @@ describe('lunaris store', function () {
           url                : '/store_insert_put',
           method             : 'PUT',
           storeName          : 'store_insert_put',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -1153,7 +1153,7 @@ describe('lunaris store', function () {
         should(lastError.length).eql(2);
         should(lastError[0]).eql('[Lunaris warn] lunaris.delete@store_del');
         should(lastError[1]).eql({ error : 404, message : 'Not Found'});
-        should(err).eql('404 : Not Found');
+        should(err).eql('${An error has occured}');
         done();
       });
 
@@ -1179,7 +1179,7 @@ describe('lunaris store', function () {
           url                : '/store_del/1',
           method             : 'DELETE',
           storeName          : 'store_del',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -1543,7 +1543,7 @@ describe('lunaris store', function () {
         should(lastError.length).eql(2);
         should(lastError[0]).eql('[Lunaris warn] lunaris.get@store');
         should(lastError[1]).eql({ error : 404, message : 'Not Found'});
-        should(err).eql('404 : Not Found');
+        should(err).eql('${An error has occured}');
         done();
       });
 
@@ -1581,7 +1581,7 @@ describe('lunaris store', function () {
           url                : '/store?limit=50&offset=0',
           method             : 'GET',
           storeName          : 'store',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -1595,7 +1595,7 @@ describe('lunaris store', function () {
       lunaris._stores['store2'] = _store;
 
       lunaris.hook('errorHttp@store2', err => {
-        should(err).eql('Error : null');
+        should(err).eql('${An error has occured}');
         done();
       });
 
@@ -2745,7 +2745,7 @@ describe('lunaris store', function () {
             url                : '/store_insert_post',
             method             : 'POST',
             storeName          : 'store_insert_post',
-            messageError       : '404 : Not Found',
+            messageError       : '${An error has occured}',
             messageErrorServer : { error : 404, message : 'Not Found'},
           });
           lunaris.retry('@store_insert_post', _values[0].url, _values[0].method, _values[0].data, _values[0].version);
@@ -2763,7 +2763,7 @@ describe('lunaris store', function () {
           url                : '/store_insert_post',
           method             : 'POST',
           storeName          : 'store_insert_post',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -2797,7 +2797,7 @@ describe('lunaris store', function () {
             url                : '/store_insert_put/2',
             method             : 'PUT',
             storeName          : 'store_insert_put',
-            messageError       : '404 : Not Found',
+            messageError       : '${An error has occured}',
             messageErrorServer : { error : 404, message : 'Not Found'},
           });
           lunaris.retry('@store_insert_put', _values[0].url, _values[0].method, _values[0].data, _values[0].version);
@@ -2814,7 +2814,7 @@ describe('lunaris store', function () {
           url                : '/store_insert_put/2',
           method             : 'PUT',
           storeName          : 'store_insert_put',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -2843,7 +2843,7 @@ describe('lunaris store', function () {
             url                : '/store?limit=50&offset=0',
             method             : 'GET',
             storeName          : 'store',
-            messageError       : '404 : Not Found',
+            messageError       : '${An error has occured}',
             messageErrorServer : { error : 404, message : 'Not Found'},
           });
           lunaris.retry('@store', _values[0].url, _values[0].method, _values[0].data, _values[0].version);
@@ -2861,7 +2861,7 @@ describe('lunaris store', function () {
           url                : '/store?limit=50&offset=0',
           method             : 'GET',
           storeName          : 'store',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();
@@ -2890,7 +2890,7 @@ describe('lunaris store', function () {
             url                : '/store_del/1',
             method             : 'DELETE',
             storeName          : 'store_del',
-            messageError       : '404 : Not Found',
+            messageError       : '${An error has occured}',
             messageErrorServer : { error : 404, message : 'Not Found'},
           });
           lunaris.retry('@store_del', _values[0].url, _values[0].method, _values[0].data, _values[0].version);
@@ -2908,7 +2908,7 @@ describe('lunaris store', function () {
           url                : '/store_del/1',
           method             : 'DELETE',
           storeName          : 'store_del',
-          messageError       : '404 : Not Found',
+          messageError       : '${An error has occured}',
           messageErrorServer : { error : 404, message : 'Not Found'},
         });
         done();